Regrab Linux DBus media keys on tab activation
Previously, if another application was opened or otherwise took up
listening for the keys, the key press events would go to that
application until closed.  This makes the behaviour consistent with
other media DBus applications in that on regaining focus, multimedia
keys will be re-grabbed and thus received by the media player's Firefox
tab.